requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.0.8
django-bootstrap4 @ file:///C:/graper_django/django-bootstrap4
django-crispy-forms==1.11.1
django-extensions==2.2.5
django-filter==2.4.0
django-mssql-backend==2.8.1
django-tables2==2.3.4
django-widget-tweaks==1.4.8
filelock==3.0.12
idna==2.10
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyodbc==4.0.30
pyOpenSSL==19.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.3.1
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==0.16.0
wfastcgi==3.0.0
